requirements.txt
aniso8601==9.0.1
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
cryptography==40.0.2
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
docx2txt==0.8
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
eureka-client==0.6.2
executing==1.2.0
Flask==2.3.1
Flask-RESTful==0.3.9
greenlet==2.0.2
idna==3.4
ifaddr==0.2.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathy==0.10.1
pdfminer.six==20221105
pdfplumber==0.9.0
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py-eureka-client==0.11.7
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.11
srsly==2.4.6
stack-data==0.6.2
thinc==8.1.9
threadpoolctl==3.1.0
tika==2.6.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Wand==0.6.11
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.3.1
widgetsnbextension==4.0.7
